Added 2D batch rendering.

+++ spectre/source/event/MouseScrollEvent.h
@@ -23,7 +23,7 @@ namespace Spectre
 			return stream.str();
 		}
 
-		int getOffset() { return m_OffsetY; }
+		double getOffset() { return m_OffsetY; }
 
 		static EventType getStaticType() { return EventType::MouseScroll; }
